Let play service deal properly with phone calls
The play service will now stop playing on an incoming phone call. If it
was playing before, it will resume playing after the call has finished.
If it was not playing before, it will not play afterwards (as it did
before). Also, notifications are built again if the play status changes
due to the call.