Catch corrupted audio file error
Previously, the app crashed when clicking on an album that contained
corrupted audio files. The reason for this was that setDataSource,
which is called in order to retrieve audio information, threw an
error.
Now, the exception is caught and the corresponding album can not be
opened.
It would be better to allow opening the album, and then prevent the
PlayActivity from opening, but right now I don't know an easy way to
do that so for now I'll leave it at that.