Fix app crash on directory removal
Previously, the app crashed when the selected directory was moved or
removed and the app was reloaded or the directory was synchronized.
Now the app will simply not show the albums anymore or show them as
deleted if the "Keep progress of deleted files in database"-option in
the settings was set to true. Fixes
https://github.com/flackbash/AudioAnchor/issues/11