Update audio_files database table in MainActivity
Previously, the audio_files table was updated only upon opening
the AlbumActivity and only for the corresponding album. Therefore, when
showing the album progress in the MainActivity, the progress for albums
that had never been clicked was shown as 00:00/00:00.
Now the audio_files table is updated in the MainActivity. To save
computation time, the update of the albums and the audio_files table is
now only performed when changing the root directory or when the user
explicitly performs a synchronization. Previously, the album table was
updated every time the app was started.