Fix import / export functionality (kinda)
Importing a database didn't work at least for some android versions
since the exported database didn't contain any tables or was otherwise
corrupted.
Now, when exporting the database, not only the .db file is exported but
also the .db-shm and .db-wal files. When importing the .db file now, the
.db-shm and .db-wal files are also imported as long as they are in the
same directory as the .db file and have the same base name.
This is a bit inconvenient, since now the user has to take care to
always move these three files together instead of only having to worry
about a single file.
Another problem is that opening the .db file with sqlite3 will remove
the .db-shm and .db-wal file. Without these files, the imported database
will either be corrupted and the app will crash or it will be empty and
the progress will be lost.