Construct paths from base directory and file names
Previously, the absolute path of an audio file was saved in the
audio_files table. Similarly, the absolute path of an album cover was
stored in the albums table.
Now, for the audio file, the path is not stored at all and instead put
together from the base directory, the album name and the audio file
name. For the album cover, the path relative to the base directory is
stored.