Show internal storage in file selector
Root privileges are required to read from certain directories. This
resulted in some directories (e.g. /storage/emulated/0) not being
displayed when traversing towards the root directory in the file
selector.
Now, when traversing towards the root directory, I store the
directories I'm coming from to be able to display them later.
This is a proper solution for #2 and #21.