Show complete texts using marquee and multi-lines
Previously, very long album titles caused a line break in the
MainActivity and the AlbumActivity. Now, the text is scrolling through
the TextView instead.
In the SettingsActivity, long settings labels were previously not
completely displayed. Now, the entire label is shown in a multi-line
TextView (for API level 26 and higher).