Enable continuation of playback in all Activities
This changes basically everything.
Previously, playback was stopped when the back button in the
PlayActivity was pressed.
Now, one can navigate through the entire app while continuing the
playback.
As long as the notification for the playback is shown (i.e. the
MediaPlayerService is running), a button at the bottom of the screen
will enable the user in the Main and AlbumActivity to resume or pause
the playback. Playback of the current audio file is stopped when the
PlayActivity is opened for a different audio file.
Previously, progress was written to the database when the back button in
the PlayActivity was pressed. Now its written to the database everytime
the playback is paused or stopped.
Many more changes were needed in order for the sleep timer and setting
changes to work properly again as well as for the Album and MainActivity
to show the current progress for the currently playing audio file or
album.