Fix app crashes in AlbumActivity for API < 21
I used an attribute to switch the color of the play status icon when
enabling dark theme. However, for API < 21, it is not allowed to use
attributes as color values for an xml drawable which caused the app
to crash.
Now I use setBackgroundResource() to switch the color of the play
status icon. This change also required adjusting the "now playing" icon
in the MainActivity.