Make settings accessible from PlayActivity
The SettingsActivity can now be accessed from within the PlayActivity.
When changing the sleep timer settings while a sleep timer is running (no
matter from where the settings are accessed), the new settings will not
take effect for the current sleep timer but only once a new sleep timer
is set.
When changing the getCoverImageFromMetadata or the getTitleFromMetadata
settings, the notification is only updated when the current playback is
paused/started.
All other setting changes should take effect immediately.
Closes #27.