Try to keep MediaPlayerService alive
This is an attempt to fix or improve #34 and #29 by e.g. using a
WakeLock and starting the service as foreground service.
This also prevents some of the playback position glitches but when the
service is killed - which unfortunately still happens - the position is
reset to the last saved position (position is saved on pause/stop and
forward/backward action).