Start service on playback start
Before, the service was started when the PlayActivity was created. Now
it is started when the playback is first started.
This entails that the notification is created on playback start.
Closes #46.
Also, call startForeground everytime the playback is started instead of
just the first time, since stopForeground is called on pause and on stop
playback. This should now once and for all fix the problem of the
service getting killed during playback and hence forgetting the last
playback position (for real this time).
Closes #34.