Add option to add two last play position bookmarks
When the option is set (disabled by default), two bookmarks are added /
updated on each play button press. One for the last play position and
one for the second to last play position in case the user only realizes
that they are at the wrong position after pressing the play button.
When the option is disabled, all last play position bookmark are
deleted.
Closes #93.