Make changes to update to target SDK version 35
This includes updated requesting and handling of permissions, as well as
using the Storage Access Framework to pick the location for a database
export or the file for import. This also required the database export to
consist of a single file. The database .db, .db-shm, and .db-wal files
are now zipped automatically into a single file upon export. Upon
import, the zipped file is automatically unzipped. This means that old
export files now need to be zipped in order to be imported again.